Typecast ObjectFlags.CreateSelected to a uint
git-svn-id: http://libopenmetaverse.googlecode.com/svn/trunk@578 52acb1d6-8a22-11de-b505-999d5b087335

+++ libsecondlife-cs/ObjectManager.cs
@@ -416,7 +416,7 @@ namespace libsecondlife
             packet.AgentData.GroupID = prim.GroupID;
 
             packet.ObjectData.State = (byte)prim.State;
-            packet.ObjectData.AddFlags = ObjectFlags.CreateSelected;
+            packet.ObjectData.AddFlags = (uint)ObjectFlags.CreateSelected;
             packet.ObjectData.PCode = (byte)PCode.Prim;
 
             packet.ObjectData.Material = (byte)prim.Material;
